Add files via upload
heart disease.csv is the dataset used in this project, it has 13 input features and 1 output feature in the ‘target’ column. i took 5 features namely cp, thal, age, oldpeak, and exang to do the diagnosis process. the output in this project is the ‘target’ column which diagnoses 1 (found symptoms of heart disease) or 0 (no indication of heart disease).